requirements.txt
plex.py==0.9.0
PlexAPI==4.15.15
texttable
python-dotenv
requests
discord.py==2.4.0
